requirements.txt
apache-airflow
apache-airflow-providers-docker
virtualenv
requests
python-binance
psycopg2-binary